requirements.txt
# Core dependencies for serving API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data handling
pandas>=2.0.0
pyarrow>=14.0.0

# ML/AI
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# Optional: for testing
requests>=2.31.0
pytest>=7.4.0
